+++ rust/src/core/stats.rs
@@ -742,7 +742,7 @@ pub fn format_gain() -> String {
     o.push(String::new());
     o.push(format!("  {DIM}{ln56}{RST}"));
     o.push(format!(
-        "  {DIM}lean-ctx v2.3.2  |  leanctx.com  |  lean-ctx dashboard{RST}"
+        "  {DIM}lean-ctx v2.3.3  |  leanctx.com  |  lean-ctx dashboard{RST}"
     ));
     o.push(String::new());