+++ rust/src/shell.rs
@@ -72,7 +72,7 @@ pub fn exec(command: &str) -> i32 {
 pub fn interactive() {
     let real_shell = detect_shell();
 
-    eprintln!("lean-ctx shell v2.3.2 (wrapping {real_shell})");
+    eprintln!("lean-ctx shell v2.3.3 (wrapping {real_shell})");
     eprintln!("All command output is automatically compressed.");
     eprintln!("Type 'exit' to quit.\n");